pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cuvis_ai"
version = "0.1.2"
description = "The CUVIS.AI Toolset"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache License 2.0" }
authors = [
    { name = "Cubert GmbH, Ulm, Germany", email = "SDK@cubert-gmbh.com" }
]

dependencies = [
"cuvis",
"setuptools; python_version>='3.12'",  # Only for Python 3.12+
"torch",
"torchvision",
"scikit-learn",
"scikit-image",
"pycocotools",
"matplotlib",
"pyyaml",
"pytorch-lightning",
"skorch",
"imantics",
"networkx",
"opencv-python",
"torchvision",
"scikit-image",
"gdown",
"dataclass-wizard"
]

classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Image Processing",
]

[project.urls]
Homepage = "https://www.cubert-hyperspectral.com/"
Repository = "https://github.com/cubert-hyperspectral/cuvis.ai"
Documentation = "https://cubert-hyperspectral.github.io/cuvis.ai/"
Changelog = "https://github.com/cubert-hyperspectral/cuvis.ai/releases"
Issues = "https://github.com/cubert-hyperspectral/cuvis.ai/issues"

[tool.setuptools]
packages = ["cuvis_ai"]
include-package-data = true